quic: implement QuicSession OnPathValidation
This commit implements path validation (Section-8.2) of the
quic-transport specification.

When an endpoint does a connection migration it needs to verify that it
can communicate with the remote peer through the new local address (the
migrated address). The migrating endpoint would send a PATH_CHALLENGE to
the remote peer. After receiving a PATH_RESPONSE successfully validation
, which is handled by ngtcp2, the local address of the connection should
be updated and used for future communication.

This commit updates the local address in the cases of successful path
validation.

What should happen if path validation fails?
We have tried to do a connection migration but failed. Should this throw and
error to indicate this. If the migration happened because a switch from wifi to
3g/4g then the old local address is probably not useable anymore. This
is left as a todo at the moment to get feedback.

Refs:
https://tools.ietf.org/html/draft-ietf-quic-transport-19#section-8.2

PR-URL: https://github.com/nodejs/quic/pull/15
Reviewed-By: James M Snell <jasnell@gmail.com>