quic: fix stateless reset
fixes sending and receiving stateless reset tokens and implement
limits.

The QUIC spec requires that stateless reset packets must be strictly
smaller than the packet that triggered it, and recommends that endpoints
keep a count of the number of stateless resets that have been sent
as a protection against endlessly looping resets and other bad things.

Also includes cleanups for QuicPacket and QuicSocket